Перенес бота в отдельный класс
