+++ .gitignore
@@ -5,3 +5,5 @@ riderModule.iml
 /_ReSharper.Caches/
 /Data/
 .idea/
+*.xml
+/.idea/
